+++ pysys-examples/internal/testcases/PySys_internal_123/Input/NestedPass/run.py
@@ -1,0 +1,10 @@
+from pysys.constants import *
+from pysys.basetest import BaseTest
+from pysys.exceptions import *
+import pysys.utils.allocport
+
+class PySysTest(BaseTest):
+	def execute(self):
+		self.log.info('Ephemeral port range is: %d-%d'%pysys.utils.allocport.getEphemeralTCPPortRange())
+	def validate(self):
+		pass